refactor(access): move groups into an expandable sidebar section
