+++ ui/src/views/admin/index.tsx
@@ -1,3 +1,4 @@
 export { default as AdminLayout } from "./AdminLayout";
 export { default as Groups } from "./Groups";
+export { default as GroupDetail } from "./GroupDetail";
 export { default as GlobalRoles } from "./GlobalRoles";
